+++ tools/font_gen/font_gen.h
@@ -23,13 +23,17 @@
 #define FONT_GEN_H
 
 #include "base/font.h"
+#include "tkc/str.h"
 #include "tkc/buffer.h"
 
 BEGIN_C_DECLS
 
 ret_t font_gen(font_t* font, uint16_t font_size, const char* str, const char* output_filename,
                const char* theme);
+/*public for test*/
 uint32_t font_gen_buff(font_t* font, uint16_t font_size, const char* str, wbuffer_t* wbuffer);
+
+const char* font_gen_expand(const char* in, str_t* out);
 
 END_C_DECLS